requirements.txt
crewai>=0.28.0
langchain>=0.0.335
openai>=1.12.0
python-dotenv>=1.0.0
requests>=2.31.0
fastapi>=0.109.0
uvicorn>=0.25.0
pydantic>=2.5.0
numpy>=1.24.3
pandas>=2.0.3
chromadb>=0.4.22
pytest>=7.4.0
black>=23.7.0
ruff>=0.1.7
flake8>=6.1.0
radon>=5.1.0 